requirements.txt
# Flask and extensions
flask==2.0.1
flask-sqlalchemy==2.5.1
flask-migrate==3.1.0
flask-cors==3.0.10
flask-wtf==1.0.0
flask-login==0.6.2  # Add this line
werkzeug==2.0.1
itsdangerous==2.0.1
jinja2==3.0.1
markupsafe==2.0.1
gunicorn==20.1.0
wtforms==3.0.0
email-validator==1.1.3

# Database
alembic==1.8.1
sqlalchemy==1.4.46
sqlcipher3-wheels>=0.4.0

# Async processing
celery==5.2.7
redis==3.5.3
click>=8.0.3,<9.0  # Compatible with both Flask and Celery

# Utilities
python-dotenv==0.19.0
pytz==2023.3
pyotp==2.9.0
pyjwt==2.10.1
qrcode==8.2
Pillow==10.0.0  # Required for QR code generation
pynacl==1.5.0  # For cryptography
# phonumbers==8.13.11  # For phone number validation and formatting
phonenumbers==9.0.9  # For phone number validation and formatting
humanize==4.7.0

# WebSocket client for WhatsApp Web
websocket-client==1.5.1
python-engineio==4.4.1
python-socketio==5.8.0

# For QR code scanning and session management
base64io==1.0.3
cryptography==40.0.2

# For browser automation (headless mode)
selenium==4.10.0
webdriver-manager==3.8.6

# For handling browser cookies and local storage
browser-cookie3==0.19.1

# For HTTP requests
requests==2.31.0

# For parsing HTML responses
beautifulsoup4==4.12.2
lxml==4.9.3

# For telegram integration (if needed)
python-telegram-bot==13.7

# Removed dependencies
# whatsapp-api-client-python - Removed as we're replacing it with our custom solution
# pywhatkit - Removed as we're replacing it
# openwa - Removed as we're replacing it
